requirements.txt
Flask==3.0.2
qrcode==7.4.2
Pillow==10.2.0
python-dotenv==1.0.1
gunicorn==21.2.0 